refactor: S2 M3 — DomainClient 5종을 Struct/Live/Test 3-way split
각 클라이언트를 다음 3개 파일로 분할:
- XxxClient.swift: struct/init + DependencyValues 접근자 + 공개 에러/도메인 타입
- XxxClient+Live.swift: liveValue (Firebase/URLSession 호출) + Live 전용 helper
- XxxClient+Test.swift: testValue (unimplemented stub)

대상: AuthClient · GoalClient · RecordClient · UserClient · OpenAIClient

분할 원칙:
- 3-way split 이후에도 공개 표면(init 시그니처, DependencyValues 접근자 이름, testValue)은 동일 유지 (M1 characterization 테스트로 회귀 방어)
- AuthClient: UIKit/GoogleSignIn/CryptoKit/randomNonceString/sha256 은 +Live 로 격리, 공개 AuthError 는 본체 유지
- OpenAIClient: loadAPIKey 는 +Live 의 fileprivate free function 으로 이동 (private static 은 파일 분리 후 접근 불가)
- Firestore 호출 코드는 현 위치 유지 (M4 에서 CoreFirebaseKit 으로 이동 예정)

검증:
- tuist generate 성공
- DomainClientTests 23건 통과 (기존 8 + characterization 15)
- Microcosm 앱 전체 빌드 성공
