+++ Projects/Domain/Client/Sources/OpenAIClient+Test.swift
@@ -1,0 +1,10 @@
+import Foundation
+import ComposableArchitecture
+import DomainEntity
+
+extension OpenAIClient: TestDependencyKey {
+    public static let testValue = OpenAIClient(
+        analyzeColor: unimplemented("\(Self.self).analyzeColor"),
+        analyzeEmotion: unimplemented("\(Self.self).analyzeEmotion")
+    )
+}
