+++ Projects/Domain/Client/Sources/RecordClient+Test.swift
@@ -1,0 +1,10 @@
+import Foundation
+import ComposableArchitecture
+import DomainEntity
+
+extension RecordClient: TestDependencyKey {
+    public static let testValue = RecordClient(
+        observe: unimplemented("\(Self.self).observe"),
+        addRecord: unimplemented("\(Self.self).addRecord")
+    )
+}
